# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Accounting & Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchase-invoices:
    get:
      operationId: listPurchaseInvoices
      tags: [purchaseInvoices]
      summary: List purchase invoices
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: partyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Purchase invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseInvoice"
    post:
      operationId: createPurchaseInvoice
      tags: [purchaseInvoices]
      summary: Create purchase invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseInvoiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseInvoice"
  /purchase-invoices/{id}:
    get:
      operationId: getPurchaseInvoice
      tags: [purchaseInvoices]
      summary: Get purchase invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseInvoice"
    patch:
      operationId: updatePurchaseInvoice
      tags: [purchaseInvoices]
      summary: Update purchase invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseInvoiceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseInvoice"
    delete:
      operationId: deletePurchaseInvoice
      tags: [purchaseInvoices]
      summary: Delete purchase invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PURCHASE ORDERS
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchaseOrders]
      summary: List purchase orders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchaseOrders]
      summary: Create purchase order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchaseOrders]
      summary: Get purchase order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    patch:
      operationId: updatePurchaseOrder
      tags: [purchaseOrders]
      summary: Update purchase order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseOrderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    delete:
      operationId: deletePurchaseOrder
      tags: [purchaseOrders]
      summary: Delete purchase order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /purchase-orders/{id}/receive:
    post:
      operationId: receivePurchaseOrder
      tags: [purchaseOrders]
      summary: Receive goods and update stock
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"

  # JOURNAL ENTRIES
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        stockItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        hsnCode:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        rate:
          type: number
        discountPct:
          type: number
        gstPct:
          type: number
        taxableAmount:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        total:
          type: number
      required:
        - id
        - itemName
        - quantity
        - unit
        - rate
        - discountPct
        - gstPct
        - taxableAmount
        - cgst
        - sgst
        - igst
        - total
    InvoiceItemInput:
      type: object
      properties:
        stockItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        hsnCode:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        rate:
          type: number
        discountPct:
          type: number
        gstPct:
          type: number
      required:
        - itemName
        - quantity
        - unit
        - rate
        - discountPct
        - gstPct
    PaymentMode:
      type: object
      properties:
        mode:
          type: string
          enum: [cash, upi, cheque, bank_transfer, credit]
        amount:
          type: number
        reference:
          type: ["string", "null"]
      required:
        - mode
        - amount
...
    PurchaseInvoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        supplierInvoiceNumber:
          type: ["string", "null"]
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        isGst:
          type: boolean
        isInterstate:
          type: boolean
        isReverseCharge:
          type: boolean
        subtotal:
          type: number
        totalTaxable:
          type: number
        totalCgst:
          type: number
        totalSgst:
          type: number
        totalIgst:
          type: number
        grandTotal:
          type: number
        amountPaid:
          type: number
        balanceDue:
          type: number
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentMode"
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - date
        - partyName
        - isGst
        - isInterstate
        - isReverseCharge
        - subtotal
        - totalTaxable
        - totalCgst
        - totalSgst
        - totalIgst
        - grandTotal
        - amountPaid
        - balanceDue
        - items
        - payments
        - createdAt
    CreatePurchaseInvoiceBody:
      type: object
      properties:
        supplierInvoiceNumber:
          type: ["string", "null"]
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        isGst:
          type: boolean
        isInterstate:
          type: boolean
        isReverseCharge:
          type: boolean
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentMode"
      required:
        - date
        - partyName
        - isGst
        - isInterstate
        - isReverseCharge
        - items
        - payments
    UpdatePurchaseInvoiceBody:
      type: object
      properties:
        supplierInvoiceNumber:
          type: ["string", "null"]
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        isGst:
          type: boolean
        isInterstate:
          type: boolean
        isReverseCharge:
          type: boolean
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentMode"
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        poNumber:
          type: string
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        status:
          type: string
          enum: [open, received, cancelled]
        grandTotal:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
      required:
        - id
        - poNumber
        - date
        - partyName
        - status
        - grandTotal
        - items
        - createdAt
    CreatePurchaseOrderBody:
      type: object
      properties:
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
      required:
        - date
        - partyName
        - items
    UpdatePurchaseOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [open, received, cancelled]
        notes:
          type: ["string", "null"]
...